bowarto/backend/requirements.txt
asgiref==3.7.2
astroid==3.0.1
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==4.3.4
cryptography==41.0.7
dill==0.3.7
Django==4.2.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-jenkins==0.110.0
django-storages==1.14.2
djangorestframework==3.14.0
entrypoints==0.3
flake8==6.1.0
idna==3.6
isodate==0.6.1
isort==5.12.0
mccabe==0.7.0
pep8==1.7.1
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pylint==3.0.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
